pyproject.toml
[tool.poetry]
name = "streamlit-tutorial"
version = "0.1.0"
description = ""
authors = ["RAJerez <rodrigoagustin23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.35.0"
pillow = "^10.3.0"
streamlit-lottie = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
